Center the content of the body
Add new declarations to the body:
set text-align to center, max-width to 600 pixels and
margin to 0 vertically and auto horizontally.

This will align the content of the body to the center
which is what we want, but it messes up with our
ingredients and steps lists.
To solve this keep ol and ul aligned to the left.

List (.links) in index.html have a different behavior
as it is expected to be at the center with no list styles.

To achieve this behavior, remove the padding and list styles and
align .links to the center.